src/main/java/com/romeao/recipebook/converters/RecipeToRecipeCommand.java
package com.romeao.recipebook.converters;

import com.romeao.recipebook.commands.RecipeCommand;
import com.romeao.recipebook.domain.Recipe;
import org.springframework.core.convert.converter.Converter;
import org.springframework.stereotype.Component;

@Component
public class RecipeToRecipeCommand implements Converter<Recipe, RecipeCommand> {
    private final NotesToNotesCommand notesConverter;
    private final IngredientToIngredientCommand ingredientConverter;
    private final CategoryToCategoryCommand categoryConverter;

    public RecipeToRecipeCommand(NotesToNotesCommand notesConverter,
                                 IngredientToIngredientCommand ingredientConverter,
                                 CategoryToCategoryCommand categoryConverter) {
        this.notesConverter = notesConverter;
        this.ingredientConverter = ingredientConverter;
        this.categoryConverter = categoryConverter;
    }

    @Override
    public RecipeCommand convert(Recipe recipe) {
        if (recipe == null) { return null; }

        RecipeCommand command = new RecipeCommand();
        command.setId(recipe.getId());
        command.setDescription(recipe.getDescription());
        command.setCookTime(recipe.getCookTime());
        command.setDifficulty(recipe.getDifficulty());
        command.setDirections(recipe.getDirections());
        command.setPrepTime(recipe.getPrepTime());
        command.setServings(recipe.getServings());
        command.setSource(recipe.getSource());
        command.setUrl(recipe.getUrl());
        command.setNotes(notesConverter.convert(recipe.getNotes()));

        recipe.getCategories()
                .forEach(cat -> command.getCategories().add(categoryConverter.convert(cat)));

        recipe.getIngredients().forEach(ingredient -> command.getIngredients()
                .add(ingredientConverter.convert(ingredient)));
        return command;
    }
}
